tools: Change from int to int64

+++ tools/typhen/typhen-api/lib/go/index.js
@@ -104,7 +104,7 @@ module.exports = function(typhen, options, helpers) {
 
     rename: function(symbol, name) {
       if (name === 'integer') {
-        return 'int';
+        return 'int64';
       } else if (name === 'void') {
         return 'nil';
       }
